Using size_t rather than unsigned

+++ src/Utils.cpp
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include "Utils.hpp"
 
-void write_png(const std::string &filename, const unsigned width, const unsigned height, std::vector<Color> &buffer) {	
+void write_png(const std::string &filename, const size_t width, const size_t height, std::vector<Color> &buffer) {	
 	// Open file for writing (binary mode)
 	FILE* fp = fopen(filename.c_str(), "wb");
 	if (fp == nullptr) {
@@ -33,8 +33,8 @@ void write_png(const std::string &filename, const unsigned width, const unsigned
 
 	// Write image data
     std::vector<png_byte> row(3 * width * sizeof(png_byte), 0); // Allocate memory for one row (3 bytes per pixel - RGB)
-	for (unsigned y = 0; y < height; ++y) {
-        for (unsigned x = 0; x < width; ++x) {
+	for (size_t y = 0; y < height; ++y) {
+        for (size_t x = 0; x < width; ++x) {
             row[x*3 + 0] = buffer[y*width + x].r;
             row[x*3 + 1] = buffer[y*width + x].g;
             row[x*3 + 2] = buffer[y*width + x].b;
